Moved all core components as a Canvas
